Move the model to its own subfolder (gesticulator/model)
